# fwber-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FWBer API Documentation",
        "description": "Comprehensive API documentation for FWBer - A proximity-based dating platform with real-time features, AI-powered matching, and privacy-first design.",
        "contact": {
            "name": "FWBer Team",
            "email": "support@fwber.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.fwber.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/chatrooms": {
            "get": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "List chatrooms",
                "description": "Get all available chatrooms with filtering",
                "operationId": "c6a9f280c28d4c2dd2c12d83631e42dd",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "interest",
                                "city",
                                "event",
                                "private"
                            ]
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "activity",
                                "newest",
                                "most_active",
                                "most_members"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated chatrooms list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Create chatroom",
                "description": "Create a new chatroom",
                "operationId": "282c14705f85d52f2ee347c648327ac1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 500
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "interest",
                                            "city",
                                            "event",
                                            "private"
                                        ]
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "neighborhood": {
                                        "type": "string"
                                    },
                                    "is_public": {
                                        "type": "boolean"
                                    },
                                    "settings": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chatrooms/{id}": {
            "get": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Get a chatroom",
                "description": "Get a specific chatroom with messages",
                "operationId": "ba11ff674ca14d485c94c2c19f10c265",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chatroom and recent messages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "chatroom": {
                                            "type": "object"
                                        },
                                        "messages": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not a member"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Update",
                "description": "Update chatroom settings (admin/moderator only)",
                "operationId": "f8618be5312c7b74f4246bc79125dead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 500
                                    },
                                    "is_public": {
                                        "type": "boolean"
                                    },
                                    "settings": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Delete",
                "description": "Delete chatroom (creator only)",
                "operationId": "bd557ea2a13f29965657a19e863291df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chatrooms/{id}/join": {
            "post": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Join",
                "description": "Join a chatroom",
                "operationId": "abbd1b76d8078f3eaa27db4e07047918",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Joined"
                    },
                    "400": {
                        "description": "Already member"
                    },
                    "403": {
                        "description": "Banned"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chatrooms/{id}/leave": {
            "post": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Leave",
                "description": "Leave a chatroom",
                "operationId": "85fabffc00ad93a1b9c0ab8a3a51b724",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Left"
                    },
                    "400": {
                        "description": "Not a member"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chatrooms/{id}/members": {
            "get": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Members",
                "description": "Get chatroom members",
                "operationId": "db9252cc388367e3be41d5a3e6ba31c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated members"
                    },
                    "403": {
                        "description": "Not a member"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chatrooms/my": {
            "get": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "My chatrooms",
                "description": "Get user's chatrooms",
                "operationId": "93c7621ecb4e83021e49764d8e599ccb",
                "responses": {
                    "200": {
                        "description": "List"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chatrooms/categories": {
            "get": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Categories",
                "description": "Get chatroom categories",
                "operationId": "73072033f79437354d610a4a78843d9f",
                "responses": {
                    "200": {
                        "description": "List"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chatrooms/popular": {
            "get": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Popular",
                "description": "Get popular chatrooms",
                "operationId": "62412b6ab7fd6834907957a9014ada8c",
                "responses": {
                    "200": {
                        "description": "List"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chatrooms/search": {
            "get": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Search",
                "description": "Search chatrooms",
                "operationId": "e74f56654e781eceae418270c3e18fdf",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "interest",
                                "city",
                                "event",
                                "private"
                            ]
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated results"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}